+++ packages/fiori-annotation-api/src/cds/document.ts
@@ -1,8 +1,18 @@
-import type { AnnotationFile, Namespace, Range, Reference } from '@sap-ux/odata-annotation-core-types';
+import { fileURLToPath } from 'url';
+
+import type { AnnotationFile, Namespace, Reference } from '@sap-ux/odata-annotation-core-types';
+import { Range } from '@sap-ux/odata-annotation-core-types';
 import type { Target } from '@sap-ux/cds-odata-annotation-converter';
 import type { Assignment, AnnotationNode } from '@sap-ux/cds-annotation-parser';
+import { type CdsCompilerFacade, type MetadataCollector, type PropagatedTargetMap } from '@sap/ux-cds-compiler-facade';
+import { toAnnotationFile, toTargetMap } from '@sap-ux/cds-odata-annotation-converter';
+import type { VocabularyService } from '@sap-ux/odata-vocabularies';
+
+import { pathFromUri } from '../utils';
+import type { TextFile } from '../types';
 
 import type { Comment } from './comments';
+import { collectComments } from './comments';
 import type { CompilerToken } from './cds-compiler-tokens';
 
 export const CDS_DOCUMENT_TYPE = 'document';
@@ -24,3 +34,125 @@ export interface Document {
 }
 
 export type AstNode = Reference | Target | Assignment | AnnotationNode | CDSDocument;
+
+/**
+ * Creates CDS document.
+ *
+ * @param serviceName - Name of the service.
+ * @param vocabularyService - Vocabulary API.
+ * @param facade - CDS compiler facade instance.
+ * @param fileCache - File content cache.
+ * @param file - File
+ * @param metadataCollector - Metadata collector instance.
+ * @returns CDS document.
+ */
+export function getDocument(
+    serviceName: string,
+    vocabularyService: VocabularyService,
+    facade: CdsCompilerFacade,
+    fileCache: Map<string, string>,
+    file: TextFile,
+    metadataCollector: MetadataCollector
+): Document {
+    const comments = getComments(fileCache, file, facade, false);
+    const cdsAnnotationFile = toTargetMap(facade.blitzIndex.forUri(file.uri), file.uri, vocabularyService, facade);
+    const tokens = facade.getTokensForUri(pathFromUri(file.uri));
+    const line = fileCache.get(file.uri)?.split('\n').length ?? 0;
+    const character = fileCache.get(file.uri)?.split('\n').pop()?.length ?? 0;
+    const range = Range.create(0, 0, line, character);
+    const cdsDocument: CDSDocument = {
+        type: 'document',
+        uri: file.uri,
+        range,
+        namespace: cdsAnnotationFile.namespace,
+        references: cdsAnnotationFile.references,
+        targets: [...(cdsAnnotationFile.targetMap || [])].map(([, value]) => value)
+    };
+
+    const annotationFile = toAnnotationFile(file.uri, vocabularyService, cdsAnnotationFile, metadataCollector).file;
+
+    filterTargets(serviceName, annotationFile);
+
+    return {
+        uri: file.uri,
+        comments,
+        ast: cdsDocument,
+        annotationFile: annotationFile,
+        tokens
+    };
+}
+
+function getComments(
+    fileCache: Map<string, string>,
+    file: TextFile,
+    facade: CdsCompilerFacade,
+    ignoreComments = true
+): Comment[] {
+    const content = fileCache.get(file.uri);
+    if (content === undefined || content === null) {
+        throw new Error(`File ${file.uri} not found in cache!`);
+    }
+    const tokenVector = facade.getTokensForUri(fileURLToPath(file.uri));
+    return ignoreComments ? [] : collectComments(tokenVector);
+}
+/**
+ * Creates ghost file document.
+ *
+ * @param serviceName - Name of the service.
+ * @param vocabularyService - Vocabulary API.
+ * @param facade - CDS compiler facade instance.
+ * @param fileCache - File content cache.
+ * @param file - File
+ * @param metadataCollector - Metadata collector instance.
+ * @param propagatedTargetMap - Propagation map.
+ * @returns Ghost file document.
+ */
+export function getGhostFileDocument(
+    serviceName: string,
+    vocabularyService: VocabularyService,
+    facade: CdsCompilerFacade,
+    fileCache: Map<string, string>,
+    file: TextFile,
+    metadataCollector: MetadataCollector,
+    propagatedTargetMap: PropagatedTargetMap
+): Document {
+    const comments = getComments(fileCache, file, facade, false);
+    const tokens = facade.getTokensForUri(pathFromUri(file.uri));
+    const cdsAnnotationFile = toTargetMap(facade.blitzIndex.forUri(file.uri), file.uri, vocabularyService, facade);
+    const annotationFile = toAnnotationFile(
+        file.uri,
+        vocabularyService,
+        cdsAnnotationFile,
+        metadataCollector,
+        undefined,
+        propagatedTargetMap
+    ).file;
+
+    filterTargets(serviceName, annotationFile);
+
+    annotationFile.uri = '!' + annotationFile.uri;
+    const cdsDocument: CDSDocument = {
+        type: 'document',
+        uri: annotationFile.uri,
+        namespace: cdsAnnotationFile.namespace,
+        references: cdsAnnotationFile.references,
+        targets: [...(cdsAnnotationFile.targetMap || [])].map(([, value]) => value)
+    };
+
+    return {
+        uri: annotationFile.uri,
+        comments,
+        tokens,
+        ast: cdsDocument,
+        annotationFile: annotationFile
+    };
+}
+
+function filterTargets(serviceName: string, annotationFile: AnnotationFile): void {
+    // only allow targets pointing to current service
+    const serviceNamespace = annotationFile.namespace?.name === serviceName ? annotationFile.namespace : undefined;
+    const aliasName = serviceNamespace ? serviceNamespace.alias : '';
+    annotationFile.targets = annotationFile.targets.filter(
+        (target) => target.name.startsWith(serviceName + '.') || (aliasName && target.name.startsWith(aliasName + '.'))
+    );
+}
